refactor parts of benchsuite/resolve.rs to a library, so it can be used by other benchmarks
